train_tangent_clip/requirements.txt
torch>=2.0.0
transformers>=4.38.0
Pillow>=9.0.0
tqdm>=4.64.0
